Revert "changed colors to u8 and changed color_edit_button_rgb to color_edit_srgb"
This reverts commit 372c7d18f814f57d40b3087a0b3e4d2e99d9ada9.

+++ src/state/brush.rs
@@ -2,9 +2,9 @@ use macroquad::prelude::*;
 use serde::{Deserialize, Serialize};
 
 pub struct Brush {
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
     pub size: f32,
     pub sw: bool,
     pub room: i32,
@@ -13,7 +13,7 @@ pub struct Brush {
     pub frame_counter : i32
 }
 impl Brush {
-    pub fn swapcolor(&self, new_color: [u8; 3]) -> Self {
+    pub fn swapcolor(&self, new_color: [f32; 3]) -> Self {
         Self {
             r: new_color[0],
             g: new_color[1],
@@ -32,8 +32,8 @@ impl Brush {
 pub struct Dot {
     pub x: f32,
     pub y: f32,
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
     pub size: f32,
 }
